The amount of oval received from the Callback_Sonar is zero

+++ examples/ros/top.hpp
@@ -7,12 +7,12 @@ using namespace ForSyDe;
 
 SC_MODULE(top)
 {
-    SY::signal<double> srca, srcb, outa, outb, outc, outd;
-    
+   //SY::signal<double> srca, srcb, outa, outb, outc, outd;
+    SY::signal<float> srca, srcb, outa, outb, outc, outd;
     SC_CTOR(top)
     {
         SY::make_roswrap("roswrap","ForsydeToRos","RosToForsyde", outa, srca); 
-        SY::make_delay("delay1", abst_ext<double>(0.0), srca, outa);
+        SY::make_delay("delay1", abst_ext<float>(0.0), srca, outa);
     }
 
 };
